% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KKalWallis.R
\name{KKalWallis}
\alias{KKalWallis}
\title{Prueba Kruskal Wallis}
\usage{
KKalWallis(x, alpha)
}
\arguments{
\item{x}{(lista) contiene los datos de los grupos}

\item{alpha}{(valor) valor signicacion}
}
\value{
Una lista con valor de H estaditico, grados de libertad, valor critico y .
}
\description{
Realiza una prueba no parametrica
}
\examples{
\dontrun{
#Primer ejemplo
datos_grupo1 <- c(14, 10, 11, 13)
datos_grupo2 <- c(16, 18, 14, 15)
datos_grupo3 <- c(16, 15, 14, 12)
datos_grupo4 <- c(17, 20, 19, 21)
datos<-list(datos_grupo1, datos_grupo2, datos_grupo3, datos_grupo4)
#Cargo la libreria
library(Kruskal.Wallis)
KKalWallis(x=datos, alpha=0.05)
#Realizar la prueba de Kruskal-Wallis
cat("Estadístico H:", resultado$chi_squared, "\n")
cat("Grados de libertad:", resultado$df, "\n")
cat("Valor crítico:", resultado$Vcritico, "\n")
}
}
